+++ commands.js
@@ -4,6 +4,7 @@ import { resetCrimsonIsleTracker } from "./features/overlays/crimsonIsleTracker"
 import { resetJerryWorkshopTracker } from "./features/overlays/jerryWorkshopTracker";
 import { resetWormMembraneProfitTracker } from "./features/overlays/wormMembraneProfitTracker";
 import { resetMagmaCoreProfitTracker } from "./features/overlays/magmaCoreProfitTracker";
+import { resetFishingProfitTracker } from "./features/overlays/fishingProfitTracker";
 
 register("command", (...args) => {
     settings.openGUI();
@@ -39,3 +40,9 @@ register("command", (...args) => {
     resetMagmaCoreProfitTracker(!!isConfirmed);
     return;
 }).setName("feeshResetMagmaCoreProfit");
+
+register("command", (...args) => {
+    const isConfirmed = args[0] && args[0] === "noconfirm";
+    resetFishingProfitTracker(!!isConfirmed);
+    return;
+}).setName("feeshResetProfitTracker");